feat(forge): move HTML shell assets to commonMain/resources
- Extracted HTML, CSS, and JS from ForgeApp.kt and ForgePersistenceScript.kt into `src/commonMain/resources/web/`.
- Included `<script src="./TrikeShed.js"></script>` in `index.html`.
- Implemented Gradle build-time generation (`generateForgeAssets`) for cross-target support (Wasm/JS environments lack runtime file access).
- Refactored generation to use chunked ByteArrays to overcome JVM bytecode "Method too large" limits.
- Validated PWA shell assembly locally via JVM, WasmJS, and JS tests.
